d3 from CDN to node packages
