+++ package.json
@@ -1,11 +1,14 @@
 {
-  "name": "test",
+  "name": "d3js",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^13.5.0",
+    "d3": "^4.0.0",
+    "d3-geo-projection": "^2.0.0",
+    "d3-scale-chromatic": "^1.0.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "5.0.0",
