+++ public/index.html
@@ -9,10 +9,6 @@
       name="description"
       content="Web site created using create-react-app"
     />
-    <script src="%PUBLIC_URL%/scripts/main.js"></script>
-    <script src="https://d3js.org/d3.v4.js"></script>
-    <script src="https://d3js.org/d3-geo-projection.v2.min.js"></script>
-    <script src="https://d3js.org/d3-scale-chromatic.v1.min.js"></script>
     <link href='http://fonts.googleapis.com/css?family=Open+Sans' rel='stylesheet' type='text/css'>
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.2.1/css/bootstrap.min.css">
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
